Allow uploading and copying images in translation suggestion (#10529)
* Updates suggestion images.

* Updates image context field.

* Address self review comments.

* Address review comments.

* Use if-else.

* Fixes lint issues.

* Fixes backend test.

* Added backend test.

* Fixes fronted test

+++ extensions/objects/templates/filepath-editor.directive.ts
@@ -353,7 +353,8 @@ angular.module('oppia').directive('filepathEditor', [
         var getTrustedResourceUrlForImageFileName = function(imageFileName) {
           if (
             ContextService.getImageSaveDestination() ===
-            IMAGE_SAVE_DESTINATION_LOCAL_STORAGE) {
+            IMAGE_SAVE_DESTINATION_LOCAL_STORAGE &&
+            ImageLocalStorageService.isInStorage(imageFileName)) {
             var imageUrl = ImageLocalStorageService.getObjectUrlForImage(
               imageFileName);
             return $sce.trustAsResourceUrl(imageUrl);
@@ -366,9 +367,12 @@ angular.module('oppia').directive('filepathEditor', [
 
         ctrl.resetFilePathEditor = function() {
           if (
-            ctrl.data.metadata.savedImageFilename &&
-            ContextService.getImageSaveDestination() ===
-            IMAGE_SAVE_DESTINATION_LOCAL_STORAGE) {
+            ctrl.data.metadata.savedImageFilename && (
+              ContextService.getImageSaveDestination() ===
+              IMAGE_SAVE_DESTINATION_LOCAL_STORAGE) &&
+            ImageLocalStorageService.isInStorage(
+              ctrl.data.metadata.savedImageFilename)
+          ) {
             ImageLocalStorageService.deleteImage(
               ctrl.data.metadata.savedImageFilename);
           }
